Chore: Separate firebase config variables
