+++ src/app/api/posts.ts
@@ -1,21 +1,11 @@
 import { Comment, Post } from '@/model/post';
 import { FullUser } from '@/model/user';
-import { initializeApp } from 'firebase/app';
-import { collection, deleteDoc, doc, getDoc, getDocs, getFirestore, orderBy, query, setDoc } from 'firebase/firestore';
-
-const firebaseConfig = {
-  apiKey: process.env.NEXT_PUBLIC_FIREBASE_API_KEY,
-  authDomain: process.env.NEXT_PUBLIC_FIREBASE_AUTH_DOMAIN,
-  databaseURL: process.env.NEXT_PUBLIC_FIREBASE_DB_URL,
-  projectId: process.env.NEXT_PUBLIC_FIREBASE_PROJECT_ID,
-};
-
-const app = initializeApp(firebaseConfig);
-const database = getFirestore(app);
+import { firebaseDB } from '@/service/firebase';
+import { collection, deleteDoc, doc, getDoc, getDocs, orderBy, query, setDoc } from 'firebase/firestore';
 
 export async function getPosts(): Promise<Post[] | []> {
   try {
-    const postsRef = collection(database, 'posts');
+    const postsRef = collection(firebaseDB, 'posts');
     const postsQuery = query(postsRef, orderBy('date', 'asc'));
     const datas = await getDocs(postsQuery);
     return Promise.all(
@@ -34,7 +24,7 @@ export async function getPosts(): Promise<Post[] | []> {
 
 export async function getPostComments(postId: string): Promise<[] | Comment[]> {
   try {
-    const commentsRef = collection(database, 'posts', postId, 'comments');
+    const commentsRef = collection(firebaseDB, 'posts', postId, 'comments');
     const commentsQuery = query(commentsRef, orderBy('createdAt', 'asc'));
     const datas = await getDocs(commentsQuery);
 
@@ -57,8 +47,8 @@ export async function getPostComments(postId: string): Promise<[] | Comment[]> {
 
 export async function addPostComment(postId: string, comment: Comment, user: FullUser) {
   try {
-    const userRef = doc(database, 'users', user.uid);
-    return await setDoc(doc(database, 'posts', postId, 'comments', comment.id), {
+    const userRef = doc(firebaseDB, 'users', user.uid);
+    return await setDoc(doc(firebaseDB, 'posts', postId, 'comments', comment.id), {
       ...comment,
       user: userRef,
     });
@@ -69,8 +59,8 @@ export async function addPostComment(postId: string, comment: Comment, user: Ful
 
 export async function updatePostComment(postId: string, comment: Comment, user: FullUser) {
   try {
-    const userRef = doc(database, 'users', user.uid);
-    return await setDoc(doc(database, 'posts', postId, 'comments', comment.id), {
+    const userRef = doc(firebaseDB, 'users', user.uid);
+    return await setDoc(doc(firebaseDB, 'posts', postId, 'comments', comment.id), {
       ...comment,
       user: userRef,
     });
@@ -80,12 +70,12 @@ export async function updatePostComment(postId: string, comment: Comment, user: 
 }
 
 export async function removePostComment(postId: string, commentId: string) {
-  return await deleteDoc(doc(database, 'posts', postId, 'comments', commentId));
+  return await deleteDoc(doc(firebaseDB, 'posts', postId, 'comments', commentId));
 }
 
 export async function getPostLikes(postId: string): Promise<[] | string[]> {
   try {
-    const likesRef = collection(database, 'posts', postId, 'likes');
+    const likesRef = collection(firebaseDB, 'posts', postId, 'likes');
     const datas = await getDocs(likesRef);
     return datas?.docs.map((doc) => doc.data()?.userId);
   } catch (error) {
@@ -98,7 +88,7 @@ export async function plusLikeCount(postId: string, userId: string) {
   try {
     const likes: string[] = await getPostLikes(postId);
     if (likes.includes(userId)) return;
-    await setDoc(doc(database, 'posts', postId, 'likes', userId), { userId });
+    await setDoc(doc(firebaseDB, 'posts', postId, 'likes', userId), { userId });
   } catch (error) {
     console.log(error);
   }
@@ -108,7 +98,7 @@ export async function minusLikeCount(postId: string, userId: string) {
   try {
     const likes: string[] = await getPostLikes(postId);
     if (!likes.includes(userId)) return;
-    await deleteDoc(doc(database, 'posts', postId, 'likes', userId));
+    await deleteDoc(doc(firebaseDB, 'posts', postId, 'likes', userId));
   } catch (error) {
     console.log(error);
   }
