+++ src/service/users.ts
@@ -1,20 +1,10 @@
 import { FullUser } from '@/model/user';
-import { initializeApp } from 'firebase/app';
-import { doc, getFirestore, setDoc } from 'firebase/firestore';
-
-const firebaseConfig = {
-  apiKey: process.env.NEXT_PUBLIC_FIREBASE_API_KEY,
-  authDomain: process.env.NEXT_PUBLIC_FIREBASE_AUTH_DOMAIN,
-  databaseURL: process.env.NEXT_PUBLIC_FIREBASE_DB_URL,
-  projectId: process.env.NEXT_PUBLIC_FIREBASE_PROJECT_ID,
-};
-
-const app = initializeApp(firebaseConfig);
-const database = getFirestore(app);
+import { doc, setDoc } from 'firebase/firestore';
+import { firebaseDB } from './firebase';
 
 export async function addUser(user: FullUser) {
   try {
-    return await setDoc(doc(database, 'users', user.uid), user);
+    return await setDoc(doc(firebaseDB, 'users', user.uid), user);
   } catch (error) {
     console.log(error);
   }
